Added Edit option Ride

+++ MyBike/MyBike/BikeScreen/View/AddBikeView.swift
@@ -83,19 +83,19 @@ struct AddBikeView: View {
     // MARK: - Private
     
     private func saveBike() {
-        var wheelSize: Double?
-        var service: Double?
+        var wheelSize: Int64?
+        var service: Int64?
         var shouldNotSave = false
         if bikeNameText.isEmpty {
             bikeNameBorder = .red
             shouldNotSave = true
         }
-        wheelSize = Double(wheelSizeText)
+        wheelSize = Int64(wheelSizeText)
         if wheelSizeText.isEmpty || wheelSize == nil {
             wheelSizeBorder = .red
             shouldNotSave = true
         }
-        service = Double(serviceInText)
+        service = Int64(serviceInText)
         if serviceInText.isEmpty || service == nil {
             serviceBorder = .red
             shouldNotSave = true
