+++ MyBike/MyBike/RideScreen/View/AddRideView.swift
@@ -126,7 +126,7 @@ struct AddRideView: View {
         newRide.bikeId = selectedBike.id
         newRide.title = textFieldTitle
         if let userDistance = distance {
-            newRide.distance = Int16(userDistance)
+            newRide.distance = Int64(userDistance)
         }
         newRide.duration = textFieldDuration
         newRide.date = textFieldDate
